Move config.js into data subdirectory
